requirements.txt
sounddevice
scipy
speechrecognition
pyttsx3
python-dotenv
google-generativeai
